requirements.txt
python-dotenv
pandas
numpy
matplotlib
scikit-learn
pymongo[srv]==3.11
pymongo
certifi

##-e .